Move non-CLI code to lib
This exposes a single `run_qemu` function. To keep this commit as simple
as possible it is a direct move of code from `main` with no changes.

The startup.nsh file is also moved to the lib in this commit. Most cargo
deps also moved to the lib.
