+++ lib/src/startup.nsh
